+++ app/src/main/java/campus/tech/kakao/map/ui/PlacesAdapter.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
