+++ app/src/main/java/campus/tech/kakao/map/viewModel/MapViewModel.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import campus.tech.kakao.map.data.MapRepository
 import campus.tech.kakao.map.model.Place
 import campus.tech.kakao.map.model.RecentSearchWord
 import com.kakao.vectormap.LatLng
