+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchActivity">
+    tools:context=".ui.SearchActivity">
 
     <LinearLayout
         android:id="@+id/searchWindow"
